+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## vNext
 
+### Chores
+
+- Update `cardano-node` to 10.1.1 via `cardano-wallet` v2024-11-18 ([PR 3229](https://github.com/input-output-hk/daedalus/pull/3229))
+
 ## 6.0.2
 
 ### Fixes